:sparkles: Implement parsing interfaces (#10)
* :refactor: Split method, interface and struct definitions

* :sparkles: Implement MethodDescriptor.Compatible

This method enables to compare two MethodDescriptor and check if they are
compatible. Method are considered compatatibles if they have the same
signature (only types of arguments and results are considered).

* :sparkles: Implement interface parsing

* :sparkles: Add `token.FSet` the `parseFileContext`

This commit enables the possibility of parseErrors be reported with
file names and line numbers.

* :bug: Composed interfaces broke the parse

This commit do not implement composing interfaces. But, it gracefully
ignores them.